+++ Assets/Amplitude/AmplitudeServerZone.cs
@@ -1,0 +1,4 @@
+public enum AmplitudeServerZone {
+    US,
+    EU
+}