+++ Assets/Plugins/iOS/Amplitude/AMPServerZoneUtil.h
@@ -1,8 +1,6 @@
 //
-//  AMPEventExplorer.h
-//  Amplitude
-//
-//  Copyright (c) 2020 Amplitude Inc. (https://amplitude.com/)
+//  AMPServerZoneUtil.h
+//  Copyright (c) 2021 Amplitude Inc. (https://amplitude.com/)
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -24,14 +22,12 @@
 //
 
 #import <Foundation/Foundation.h>
+#import "AMPServerZone.h"
 
-NS_ASSUME_NONNULL_BEGIN
+@interface AMPServerZoneUtil : NSObject
 
-@interface AMPEventExplorer : NSObject
++ (NSString *)getEventLogApi:(AMPServerZone)serverZone;
 
-- (instancetype)initWithInstanceName:(NSString *)instanceName;
-- (void)showBubbleView;
++ (NSString *)getDynamicConfigApi:(AMPServerZone)serverZone;
 
 @end
-
-NS_ASSUME_NONNULL_END