+++ Assets/Plugins/iOS/Amplitude/Amplitude.h
@@ -25,12 +25,14 @@
 #import "AMPIdentify.h"
 #import "AMPRevenue.h"
 #import "AMPTrackingOptions.h"
+#import "AMPPlan.h"
+#import "AMPServerZone.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
 typedef NSString *_Nonnull (^AMPAdSupportBlock)(void);
 typedef NSDictionary *_Nullable (^AMPLocationInfoBlock)(void);
-
+typedef void (^AMPInitCompletionBlock)(void);
 /**
  Amplitude iOS SDK.
 
@@ -108,7 +110,7 @@ typedef NSDictionary *_Nullable (^AMPLocationInfoBlock)(void);
 @property (nonatomic, assign) int eventUploadMaxBatchSize;
 
 /**
- The maximum number of events that can be stored lcoally. The default is 1000 events.
+ The maximum number of events that can be stored locally. The default is 1000 events.
  */
 @property (nonatomic, assign) int eventMaxCount;
 
@@ -176,13 +178,21 @@ typedef NSDictionary *_Nullable (^AMPLocationInfoBlock)(void);
  */
 @property (nonatomic, strong, nullable) AMPLocationInfoBlock locationInfoBlock;
 
-#if TARGET_OS_IOS || TARGET_OS_MACCATALYST
 /**
- Show Amplitude Event Explorer when you're running a debug build.
+ Content-Type header for event sending requests. Only relevant for sending events to a different URL (e.g. proxy server)
  */
-@property (nonatomic, assign, readwrite) BOOL showEventExplorer;
+@property (nonatomic, copy, readonly) NSString *contentTypeHeader;
 
-#endif
+/**
+ * Sets a block to be called after completely initialized.
+ *
+ * Example:
+ *  __typeof(amp) __weak weakAmp = amp;
+ *  amp.initCompletionBlock = ^(void){
+ *     NSLog(@"deviceId: %@, userId: %@", weakAmp.deviceId, weakAmp.userId);
+ *  };
+ */
+@property (nonatomic, strong, nullable) AMPInitCompletionBlock initCompletionBlock;
 
 #pragma mark - Methods
 
@@ -594,6 +604,14 @@ typedef NSDictionary *_Nullable (^AMPLocationInfoBlock)(void);
  */
 - (void)useAdvertisingIdForDeviceId;
 
+/**
+  By default the iOS SDK will track several user properties such as carrier, city, country, ip_address, language, platform, etc. You can use the provided AMPTrackingOptions interface to customize and disable individual fields.
+
+  Note: Each operation on the AMPTrackingOptions object returns the same instance which allows you to chain multiple operations together.
+
+      AMPTrackingOptions *options = [[[[AMPTrackingOptions options] disableCity] disableIPAddress] disablePlatform];
+      [[Amplitude instance] setTrackingOptions:options];
+ */
 - (void)setTrackingOptions:(AMPTrackingOptions *)options;
 
 /**
@@ -607,9 +625,34 @@ typedef NSDictionary *_Nullable (^AMPLocationInfoBlock)(void);
  */
 - (void)disableCoppaControl;
 
+/**
+ Sends events to a different URL other than kAMPEventLogUrl. Used for proxy servers
+ 
+ We now have a new method setServerZone. To send data to Amplitude's EU servers, recommend to use setServerZone
+ method like [client setServerZone:EU]
+ */
 - (void)setServerUrl:(NSString *)serverUrl;
 
+/**
+ Sets Content-Type header for event sending requests
+*/
+- (void)setContentTypeHeader:(NSString *)contentType;
+
 - (void)setBearerToken:(NSString *)token;
+
+- (void)setPlan:(AMPPlan *)plan;
+
+/**
+ * Set Amplitude Server Zone, switch to zone related configuration, including dynamic configuration and server url.
+ * To send data to Amplitude's EU servers, you need to configure the serverZone to EU like [client setServerZone:EU]
+ */
+- (void)setServerZone:(AMPServerZone)serverZone;
+
+/**
+ * Set Amplitude Server Zone, switch to zone related configuration, including dynamic configuration and server url.
+ * If updateServerUrl is true, including server url as well. Recommend to keep updateServerUrl to be true for alignment.
+ */
+- (void)setServerZone:(AMPServerZone)serverZone updateServerUrl:(BOOL)updateServerUrl;
 
 /**-----------------------------------------------------------------------------
  * @name Other Methods
@@ -668,6 +711,9 @@ typedef NSDictionary *_Nullable (^AMPLocationInfoBlock)(void);
  Call to check if the SDK is ready to start a new session at timestamp. Returns YES if a new session was started, otherwise NO and current session is extended. Only use if you know what you are doing. Recommended to use current time in UTC milliseconds for timestamp.
  */
 - (BOOL)startOrContinueSession:(long long)timestamp;
+
+
+- (NSString *)getContentTypeHeader;
 
 @end